+++ src/matchers/Ip.js
@@ -7,11 +7,19 @@
 import UrlMatcher from './Url';
 import { IP_PATTERN } from '../constants';
 
-import type { MatchResponse } from '../types';
+import type { MatchResponse, MatcherFactory } from '../types';
 
 const IP_REGEX = new RegExp(IP_PATTERN, 'i');
 
 export default class IpMatcher extends UrlMatcher {
+  constructor(name: string, options: Object = {}, factory: ?MatcherFactory = null) {
+    super(name, {
+      ...options,
+      // IPs dont have TLDs
+      validateTLD: false,
+    }, factory);
+  }
+
   match(string: string): ?MatchResponse {
     return this.doMatch(string, IP_REGEX, this.handleMatches);
   }